refactor of controllers and commands packaged and other minors
