+++ app/src/main/java/org/WenuLink/adapters/aircraft/TelemetryData.kt
@@ -1,6 +1,5 @@
 package org.WenuLink.adapters.aircraft
 
-import com.MAVLink.enums.CAMERA_MODE
 import dji.common.remotecontroller.HardwareState.FlightModeSwitch
 import kotlin.Int
 import kotlin.math.roundToInt
