+++ app/src/main/java/org/WenuLink/adapters/aircraft/TelemetryHandler.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.flow.distinctUntilChangedBy
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.WenuLink.adapters.AsyncUtils
-import org.WenuLink.adapters.commands.IHandler
+import org.WenuLink.commands.IHandler
 import org.WenuLink.sdk.AircraftManager
 import org.WenuLink.sdk.FCManager
 import org.WenuLink.sdk.RCManager
