+++ app/src/main/java/org/WenuLink/commands/IHandler.kt
@@ -1,4 +1,4 @@
-package org.WenuLink.adapters.commands
+package org.WenuLink.commands
 
 import kotlinx.coroutines.CoroutineScope
